camel-fhir record FHIR context instead of creating it at runtime. Only one FHIR context per specification is permitted in native mode
