+++ integration-tests/fhir/src/main/resources/application.properties
@@ -37,3 +37,4 @@ quarkus.camel.fhir.enable-r5=false
 camel.component.fhir.configuration = #class:org.apache.camel.component.fhir.FhirConfiguration
 camel.component.fhir.configuration.log = false
 camel.component.fhir.configuration.server-url = {{camel.fhir.test-url}}
+camel.component.fhir.configuration.fhir-context = #bean:DSTU3
